Add a Builder API to create Clauses in a Formula
The creation of clauses is not very readable with the
current createClause API. One had to first create the
vector of RawLiterals, then add them one by one and
finally call formula->createClause().

So a new ClauseBuilder class enables one to write:
  formula.newClause(clauseId)
      .withPositiveLiteral(1)
      .withPositiveLiteral(3)
      .withNegativeLiteral(4)
      .build();

This is a lot more concise, readable and less error-prone.