Break the shared_ptr cycle between Clause and Variable
A Clause references Variables through Literals, and Variables
references Clauses in which they appear.

All those references were handled using shared_ptr, which led
to cycles and indirect memory leaks.

In order to break the cycles, the references from a Literal to
its associated Variable is no done through a weak_ptr.
This is enough to break the cycle and remove the indirect memory
leaks reported by libasan.